+++ src/main/java/at/htlkaindorf/clashtoolsbackend/dto/account/AccountEntityRequestDTO.java
@@ -35,5 +35,5 @@ public class AccountEntityRequestDTO {
      * The timestamp when the upgrade for this entity started, if any.
      * This field is optional and can be null if no upgrade is in progress.
      */
-    private Integer upgradeStart;
+    private Long upgradeStart;
 }
